Refactor text response solutions for reuse
