+++ client/app/bundles/course/assessment/submission/components/answers/TextResponse.jsx
@@ -1,181 +1,12 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { Field } from 'redux-form';
-import { defineMessages, FormattedMessage } from 'react-intl';
 import RichTextField from 'lib/components/redux-form/RichTextField';
-import { Table, TableBody, TableHeader, TableHeaderColumn,
-  TableRow, TableRowColumn } from 'material-ui/Table';
 
 import { questionShape } from '../../propTypes';
 import UploadedFileView from '../../containers/UploadedFileView';
+import TextResponseSolutions from '../TextResponseSolutions';
 import FileInput from '../FileInput';
-
-const translations = defineMessages({
-  solutions: {
-    id: 'course.assessment.submission.answer.solutions',
-    defaultMessage: 'Solutions',
-  },
-  solutionsWithMaximumGrade: {
-    id: 'course.assessment.submission.answer.solutionsWithMaximumGrade',
-    defaultMessage: 'Solutions (Maximum Grade for this Question: {maximumGrade})',
-  },
-  type: {
-    id: 'course.assessment.submission.answer.type',
-    defaultMessage: 'Type',
-  },
-  solution: {
-    id: 'course.assessment.submission.answer.solution',
-    defaultMessage: 'Solution',
-  },
-  information: {
-    id: 'course.assessment.submission.answer.information',
-    defaultMessage: 'Word from Text Passage',
-  },
-  grade: {
-    id: 'course.assessment.submission.answer.grade',
-    defaultMessage: 'Grade',
-  },
-  group: {
-    id: 'course.assessment.submission.answer.group',
-    defaultMessage: 'Group',
-  },
-  point: {
-    id: 'course.assessment.submission.answer.point',
-    defaultMessage: 'Point',
-  },
-  maximumGroupGrade: {
-    id: 'course.assessment.submission.answer.maximumGroupGrade',
-    defaultMessage: 'Maximum Grade for this Group',
-  },
-  pointGrade: {
-    id: 'course.assessment.submission.answer.pointGrade',
-    defaultMessage: 'Grade for this Point',
-  },
-  compre_keyword: {
-    id: 'course.assessment.submission.answer.compreKeyword',
-    defaultMessage: 'Keywords',
-  },
-  compre_lifted_word: {
-    id: 'course.assessment.submission.answer.compreLiftedWord',
-    defaultMessage: 'Lifted Words',
-  },
-  solutionLemma: {
-    id: 'course.assessment.submission.answer.solutionLemma',
-    defaultMessage: 'Solution (lemma form for autograding)',
-  },
-});
-
-function renderTextResponseSolutions(question) {
-  return (
-    <>
-      <hr />
-      <h4><FormattedMessage {...translations.solutions} /></h4>
-      <Table selectable={false}>
-        <TableHeader adjustForCheckbox={false} displaySelectAll={false}>
-          <TableRow>
-            <TableHeaderColumn><FormattedMessage {...translations.type} /></TableHeaderColumn>
-            <TableHeaderColumn><FormattedMessage {...translations.solution} /></TableHeaderColumn>
-            <TableHeaderColumn><FormattedMessage {...translations.grade} /></TableHeaderColumn>
-          </TableRow>
-        </TableHeader>
-        <TableBody displayRowCheckbox={false}>
-          {question.solutions.map(solution => (
-            <TableRow key={solution.id}>
-              <TableRowColumn>{solution.solutionType}</TableRowColumn>
-              <TableRowColumn style={{ whiteSpace: 'pre-wrap' }}>{solution.solution}</TableRowColumn>
-              <TableRowColumn>{solution.grade}</TableRowColumn>
-            </TableRow>
-          ))}
-        </TableBody>
-      </Table>
-    </>
-  );
-}
-
-function renderTextResponseComprehensionPoint(point) {
-  return (
-    <>
-      <br />
-      <h5><FormattedMessage {...translations.point} /></h5>
-      <Table selectable={false}>
-        <TableBody displayRowCheckbox={false}>
-          <TableRow>
-            <TableRowColumn><FormattedMessage {...translations.pointGrade} /></TableRowColumn>
-            <TableRowColumn>{point.pointGrade}</TableRowColumn>
-            <TableRowColumn />
-            <TableRowColumn />
-          </TableRow>
-          <TableRow>
-            <TableHeaderColumn><FormattedMessage {...translations.type} /></TableHeaderColumn>
-            <TableHeaderColumn><FormattedMessage {...translations.solution} /></TableHeaderColumn>
-            <TableHeaderColumn><FormattedMessage {...translations.solutionLemma} /></TableHeaderColumn>
-            <TableHeaderColumn><FormattedMessage {...translations.information} /></TableHeaderColumn>
-          </TableRow>
-          {point.solutions.map(solution => (
-            <TableRow key={solution.id}>
-              <TableRowColumn><FormattedMessage {...translations[solution.solutionType]} /></TableRowColumn>
-              <TableRowColumn style={{ whiteSpace: 'pre-wrap' }}>{solution.solution}</TableRowColumn>
-              <TableRowColumn style={{ whiteSpace: 'pre-wrap' }}>{solution.solutionLemma}</TableRowColumn>
-              <TableRowColumn style={{ whiteSpace: 'pre-wrap' }}>{solution.information}</TableRowColumn>
-            </TableRow>
-          ))}
-        </TableBody>
-      </Table>
-    </>
-  );
-}
-
-function renderTextResponseComprehensionGroup(group) {
-  return (
-    <>
-      <br />
-      <h4><FormattedMessage {...translations.group} /></h4>
-      <Table selectable={false}>
-        <TableBody displayRowCheckbox={false}>
-          <TableRow>
-            <TableRowColumn><FormattedMessage {...translations.maximumGroupGrade} /></TableRowColumn>
-            <TableRowColumn>{group.maximumGroupGrade}</TableRowColumn>
-            <TableRowColumn />
-            <TableRowColumn />
-          </TableRow>
-          {group.points.map(point => (
-            <TableRow key={point.id}>
-              <TableRowColumn colSpan={4}>
-                {renderTextResponseComprehensionPoint(point)}
-              </TableRowColumn>
-            </TableRow>
-          ))}
-        </TableBody>
-      </Table>
-    </>
-  );
-}
-
-function renderTextResponseComprehension(question) {
-  return (
-    <>
-      <hr />
-      <h4>
-        <FormattedMessage
-          {...translations.solutionsWithMaximumGrade}
-          values={{ maximumGrade: question.maximumGrade }}
-        />
-      </h4>
-      {question.groups.map(group => (
-        <div key={group.id}>{renderTextResponseComprehensionGroup(group)}</div>
-      ))}
-    </>
-  );
-}
-
-function solutionsTable(question) {
-  if (question.comprehension && question.groups) {
-    return renderTextResponseComprehension(question);
-  } if (!question.comprehension && question.solutions) {
-    return renderTextResponseSolutions(question);
-  }
-  return null;
-}
 
 function TextResponse({ question, readOnly, answerId, graderView }) {
   const allowUpload = question.allowAttachment;
